Change implementation
Instead of calling CodeBuild and CodePipeline APIs per request,
capture all relavant CB and CP events, and store project state
in a DynamoDB table. Use that state when responding to CCTray request.

This means a less direct solution but I've done this because
the direct APIs require a lot of work that I hadn't yet implemented,
could in theory get throttled for accounts with a high
number of projects, and also would be starting to get slow for a high number of projects.